added environment variables to imports()
