+++ conans/client/importer.py
@@ -8,6 +8,7 @@ from conans.util.files import save, md5sum, load
 import calendar
 import time
 from conans.errors import ConanException
+from conans.tools import environment_append
 
 
 IMPORTS_MANIFESTS = "conan_imports_manifest.txt"
@@ -50,7 +51,8 @@ def undo_imports(current_path, output):
 def run_imports(conanfile, current_path, output):
     file_importer = FileImporter(conanfile, current_path)
     conanfile.copy = file_importer
-    conanfile.imports()
+    with environment_append(conanfile.env):
+        conanfile.imports()
     copied_files = file_importer.execute()
     import_output = ScopedOutput("%s imports()" % output.scope, output)
     report_copied_files(copied_files, import_output)
